backend/server_alpha/requirements.txt
django==2.0.5
psycopg2==2.7.4
django-oauth-toolkit==1.0.0
django-cors-middleware==1.3.1
djangorestframework==3.7.7
python-decouple==3.1
dj-database-url==0.5.0
django-filter==1.1.0
rest_condition==1.0.3
google-api-python-client==1.6.6
google-python-cloud-debugger==2.8
google-auth==1.4.1
google-auth-oauthlib==0.2.0
google-auth-httplib2==0.0.3
google-cloud-storage==1.10.0
django-rest-swagger==2.2.0
django-storages==1.6.6
boto==2.48.0
